Reduced code duplication
